// src/app/Models/authenticationRes.ts
export class authenticationRes {
  'id': number;
  'username': string;
  'email': string;
  'email_verified_at': string;
  'profilePicturePath': string;
  'address': string;
  'phoneNumber': number;
  'isAdmin': number;
  'isSuperAdmin': number;
  'created_at': string;
  'updated_at': string;
  'token': string;
}